third_party/WebKit/LayoutTests/fast/canvas/canvas-overloads-strokeText-expected.txt

Test the behavior of CanvasRenderingContext2D.strokeText() when called with different numbers of arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.strokeText() threw exception TypeError: Failed to execute 'strokeText' on 'CanvasRenderingContext2D': 3 arguments required, but only 0 present..
PASS ctx.strokeText('moo') threw exception TypeError: Failed to execute 'strokeText' on 'CanvasRenderingContext2D': 3 arguments required, but only 1 present..
PASS ctx.strokeText('moo',0) threw exception TypeError: Failed to execute 'strokeText' on 'CanvasRenderingContext2D': 3 arguments required, but only 2 present..
PASS ctx.strokeText('moo',0,0) is undefined
PASS ctx.strokeText('moo',0,0,0) is undefined
PASS ctx.strokeText('moo',0,0,0,0) is undefined
PASS successfullyParsed is true

TEST COMPLETE
